Refactoring with Clap enums
