+++ Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-clap = { version = "4.4.4", features = ["cargo"] }
+clap = { version = "4.4.4", features = ["cargo", "derive"] }
 colored = "2.0.4"
 compare = "0.1.0"
 csv = "1.3.0"
